catch potential issue with deleting files

+++ APIs/GetGameList.php
@@ -236,9 +236,11 @@ function deleteDirectory($dir)
 
   if (!is_dir($dir)) {
     $handler = fopen($dir, "w");
-    fwrite($handler, "");
-    fclose($handler);
-    return unlink($dir);
+    if ($handler) {
+      fwrite($handler, "");
+      fclose($handler);
+      return unlink($dir);
+    }
   }
 
   foreach (scandir($dir) as $item) {
